requirements.txt
streamlit
requests
nltk
gtts
python-dotenv